First implementation of Match Pool resource
It works, kind of, but we need to understand some of the API return values better. Some are always marked as changed (false -> null) for example.

Perhaps there's a chance to reduce the amount of repetition between Create & Update calls?

Perhaps there's a chance to reduce the amount of repetition after Read operations?

Should we update our internal state with the response from the Update operation?